+++ src/hooks/index.ts
@@ -1,1 +1,2 @@
 export { useMergedFocusHandlers } from './useMergedFocusHandlers'
+export { useSamenessNumberState } from './useSamenessNumberState'